LayoutTests/fast/events/popup-blocked-from-fake-button-click.html

<html> <!-- webkit-test-runner [ JavaScriptCanOpenWindowsAutomatically=false ] -->
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <body>
        <button id="test" onclick="popup()" style="display:none"></button>
        <div id="console"></div>
        <script> 
            window.jsTestIsAsync = true;
            var win;
            function popup() {
                win = window.open("about:blank", "blank");
                shouldBeNull("win");
            }


            document.getElementById("test").click();

            finishJSTest();
        </script> 
    </body> 
</html>
